+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @primer/brand-storybook
 
+## 0.31.0
+
 ## 0.30.3
 
 ## 0.30.2